requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==6.0.0
httplib2==0.22.0
idna==3.6
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.1.4
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
StrEnum==0.4.15
tabulate==0.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.5
python-dotenv==0.20.0
boto3
slack_sdk